Move diffusion generate routing helpers
