+++ python/sglang/cli/generate.py
@@ -3,30 +3,6 @@ import argparse
 from sglang.cli.utils import get_is_diffusion_model, get_model_path
 
 
-def _get_pipeline_class_name(extra_argv):
-    for i, arg in enumerate(extra_argv):
-        if arg == "--pipeline-class-name" and i + 1 < len(extra_argv):
-            return extra_argv[i + 1]
-        if arg.startswith("--pipeline-class-name="):
-            return arg.split("=", 1)[1]
-    return None
-
-
-def _has_registered_pipeline_class(extra_argv):
-    pipeline_class_name = _get_pipeline_class_name(extra_argv)
-    if not pipeline_class_name:
-        return False
-    try:
-        from sglang.multimodal_gen.registry import (
-            _PIPELINE_REGISTRY,
-            _discover_and_register_pipelines,
-        )
-    except ImportError:
-        return False
-    _discover_and_register_pipelines()
-    return pipeline_class_name in _PIPELINE_REGISTRY
-
-
 def generate(args, extra_argv):
     # If help is requested, show generate subcommand help without requiring --model-path
     if any(h in extra_argv for h in ("-h", "--help")):
@@ -40,9 +16,13 @@ def generate(args, extra_argv):
         return
 
     model_path = get_model_path(extra_argv)
-    is_diffusion_model = get_is_diffusion_model(
-        model_path
-    ) or _has_registered_pipeline_class(extra_argv)
+    is_diffusion_model = get_is_diffusion_model(model_path)
+    if not is_diffusion_model:
+        from sglang.multimodal_gen.runtime.entrypoints.cli.generate_routing import (
+            _has_registered_pipeline_class,
+        )
+
+        is_diffusion_model = _has_registered_pipeline_class(extra_argv)
     if is_diffusion_model:
         from sglang.multimodal_gen.runtime.entrypoints.cli.generate import (
             add_multimodal_gen_generate_args,
